// face-check-ui/src/app/services/models/location-record-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface LocationRecordDto {
  accuracy?: number;
  altitude?: number;
  batteryLevel?: number;
  bearing?: number;
  distanceFromPrevious?: number;
  id?: number;
  latitude?: number;
  longitude?: number;
  speed?: number;
  timestamp?: string;
  userFirstName?: string;
  userId?: number;
  userLastName?: string;
}